third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-late-expected.txt

CONSOLE ERROR: line 25: Uncaught SyntaxError: Unexpected token }
This is a testharness.js-based test.
FAIL Invalid uncompiled raw handlers should only be compiled when about to call them. assert_array_equals: property 1, expected "error" but got "Error"
Harness: the test ran to completion.
